Implemented an open PlayerServiceViewModel class
