+++ app/src/main/java/com/aaa/vibesmusic/ui/screens/playlist/PlaylistScreenViewModel.kt
@@ -2,10 +2,7 @@ package com.aaa.vibesmusic.ui.screens.playlist
 
 import android.Manifest
 import android.app.Application
-import android.content.ComponentName
 import android.content.Context
-import android.content.ServiceConnection
-import android.os.IBinder
 import androidx.activity.compose.ManagedActivityResultLauncher
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
@@ -13,7 +10,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
@@ -24,10 +20,9 @@ import com.aaa.vibesmusic.database.VibesMusicDatabase
 import com.aaa.vibesmusic.database.data.music.Song
 import com.aaa.vibesmusic.database.data.playlist.PlaylistSongs
 import com.aaa.vibesmusic.perms.PermissionsUtil
-import com.aaa.vibesmusic.player.MediaPlayerService
-import java.util.Objects
+import com.aaa.vibesmusic.ui.viewmodel.PlayerServiceViewModel
 
-class PlaylistScreenViewModel(application: Application, playlistId: Int) : AndroidViewModel(application) {
+class PlaylistScreenViewModel(application: Application, playlistId: Int) : PlayerServiceViewModel(application) {
     companion object {
         fun getFactory(playlistId: Int): ViewModelProvider.Factory{
             return viewModelFactory {
@@ -40,20 +35,6 @@ class PlaylistScreenViewModel(application: Application, playlistId: Int) : Andro
 
     private val db: VibesMusicDatabase = VibesMusicDatabase.getInstance(this.getApplication())
 
-    // Player service
-    private var playerService: MediaPlayerService? by mutableStateOf(null)
-    private val serviceConnection: ServiceConnection = object : ServiceConnection {
-        override fun onServiceConnected(name: ComponentName?, service: IBinder?) {
-            val binder: MediaPlayerService.MediaPlayerServiceBinder = service as MediaPlayerService.MediaPlayerServiceBinder
-            playerService = binder.mediaPlayerService
-        }
-
-        override fun onServiceDisconnected(name: ComponentName?) {
-            playerService = null
-        }
-
-    }
-
     // Playlist songs state
     var playlistSongs: PlaylistSongs? by mutableStateOf(null)
     private var playlistSongsLiveData: LiveData<PlaylistSongs> = this.getPlaylistSongsLiveData(playlistId)
@@ -63,13 +44,12 @@ class PlaylistScreenViewModel(application: Application, playlistId: Int) : Andro
 
     init {
         this.playlistSongsLiveData.observeForever(this.playlistSongObserver)
-        this.initPlayerService()
     }
 
     fun onSongClicked(launcher: ManagedActivityResultLauncher<String, Boolean>, context: Context, index: Int) {
         if(!PermissionsUtil.hasPermission(context, Manifest.permission.POST_NOTIFICATIONS))
             launcher.launch(Manifest.permission.POST_NOTIFICATIONS)
-        this.playerService?.setSongs(this.getPlaylistSongs(), index)
+        super.playerService?.setSongs(this.getPlaylistSongs(), index)
     }
 
     fun getPlaylistSongs(): List<Song> {
@@ -79,15 +59,9 @@ class PlaylistScreenViewModel(application: Application, playlistId: Int) : Andro
     @Composable
     fun getNotificationsPermissionLauncher(): ManagedActivityResultLauncher<String, Boolean> {
         return rememberLauncherForActivityResult(contract = ActivityResultContracts.RequestPermission()) { isGranted ->
-            if(playerService?.isPlaying == true && isGranted) {
-                playerService?.showNotification()
+            if(super.playerService?.isPlaying == true && isGranted) {
+                super.playerService?.showNotification()
             }
-        }
-    }
-
-    private fun initPlayerService() {
-        this.playerService ?: run {
-            MediaPlayerService.bindTo(this.getApplication(), this.serviceConnection)
         }
     }
 
@@ -98,6 +72,5 @@ class PlaylistScreenViewModel(application: Application, playlistId: Int) : Andro
     override fun onCleared() {
         super.onCleared()
         this.playlistSongsLiveData.removeObserver(this.playlistSongObserver)
-        getApplication<Application>().unbindService(this.serviceConnection)
     }
 }
